refactor: view 코드 리팩터링 InputView, KoreanInputView -> InputView로 통합
