+++ src/main/java/racingcar/ConsoleApplication.java
@@ -4,18 +4,16 @@ import racingcar.controller.RacingConsoleController;
 import racingcar.exception.CustomException;
 import racingcar.model.car.strategy.MovingStrategy;
 import racingcar.model.car.strategy.RandomMovingStrategy;
-import racingcar.view.inputview.InputView;
-import racingcar.view.inputview.KoreanInputView;
-import racingcar.view.outputview.KoreanOutputView;
-import racingcar.view.outputview.OutputView;
+import racingcar.view.InputView;
+import racingcar.view.OutputView;
 
 import java.util.Scanner;
 
 public class ConsoleApplication {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        InputView inputView = new KoreanInputView(scanner);
-        OutputView outputView = new KoreanOutputView();
+        InputView inputView = new InputView(scanner);
+        OutputView outputView = new OutputView();
         MovingStrategy movingStrategy = new RandomMovingStrategy();
 
         RacingConsoleController racingConsoleController
