+++ src/main/java/racingcar/view/InputView.java
@@ -1,22 +1,22 @@
-package racingcar.view.inputview;
+package racingcar.view;
 
 import java.util.Scanner;
 
-public class KoreanInputView extends InputView {
+public class InputView {
 
-    public KoreanInputView(Scanner scanner) {
-        super(scanner);
+    private Scanner scanner;
+
+    public InputView(final Scanner scanner) {
+        this.scanner = scanner;
     }
 
-    @Override
     public String inputCarNames() {
         System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
-        return super.nextLine();
+        return scanner.nextLine().strip();
     }
 
-    @Override
     public String inputTrialTimes() {
         System.out.println("시도할 횟수는 몇회인가요?");
-        return super.nextLine();
+        return scanner.nextLine().strip();
     }
 }
